requirements.txt
django>=3.2.4,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2-binary>=2.8,<2.9
transliterate>=1.10.2,<1.11.0
